// src/types/uauTypes.ts
export type PessoaUauRequest = {
  CodigoPessoa: number;
  NomePessoa: string;
  TipoPessoa: number;
  CpfPessoa: string;
  DataCadastroPessoa: string;
  DataNascimentoPessoa: string | null;
  InternoExterno: number;
  UsuarioCadastro: string;
  UsuarioAlteracao: string;
  StatusPessoa: number;
  TratamentoPessoa: string;
  SiglaObraPessoa: string | null;
  EmailPessoa: string;
  EnderecoEletronicoPessoa: string;
  MatriculaPessoa: string | null;
  AtivoInativo: number;
  DataAlteracaoPessoa: string;
  NomeFantasia: string;
  AnexosPessoa: number;
  InscricaoMunicipal: string;
  InscricaoEstadual: string;
  SiglaEmpresaPessoa: string | null;
  LoginPessoa: string;
  SenhaLoginPessoa: string;
  CnaePessoa: string | null;
  DataCadastroPortal: string | null;
  PessoaCadastradoPrefeituraGyn: boolean;
  HabilitarRiscoSacado: boolean;
  CeiPessoa: string | null;
  ListaDadosPessoaJuridica: any[]; // Array vazio no exemplo, ajuste se houver interface
  ListaDadosPessoFisica: DadosPessoaFisica[];
  ListaDadosPessoaEndereco: DadosPessoaEndereco[];
};

type DadosPessoaFisica = {
  CodigoPessoa: number;
  LotacaoPessoa: string;
  CargoPessoa: string;
  DataAdmissao: string | null;
  EnderecoCorrespondencia: number;
  EstadoCivil: number;
  DocumentoPessoa: string;
  TipoDocumento: string;
  DataDocumento: string | null;
  SexoPessoa: number;
  NacionalidadePessoa: string;
  NumeroDependentes: number;
  NomePai: string;
  DataNascimentoPai: string;
  NomeMae: string;
  DataNascimentoMae: string;
  NaturalidadePessoa: string;
  CodigoNacao: number | null;
  CodigoGrauInstrucao: string;
  CodigoRaca: string;
  CodigoSituacaoMilitar: string | null;
  UfNascimento: string | null;
  CidadeNatal: number;
  FatorRH: string | null;
  RegistroCasamento: string | null;
  NumeroCdi: string | null;
  NumeroProfissao: string | null;
  UniaoEstavel: boolean;
  DetalhaNacao: string | null;
  CodigoNacaoOrigem: string | null;
  CodigoMunicipalNascimento: string | null;
  NumeroIndicativoFiscal: string | null;
  CodigoFormaTributacao: string | null;
};

type DadosPessoaEndereco = {
  CodigoPessoa: number;
  TipoEndereco: number;
  EnderecoPessoa: string | null;
  BairroPessoa: string | null;
  CidadePessoa: string | null;
  UfPessoa: string | null;
  CepPessoa: string | null;
  NumeroEnderecoPessoa: string | null;
  ComplementoEndereco: string | null;
  ReferenciaEndereco: string | null;
  EnderecoProprio: number | null;
  NumeroCidadeEndereco: number | null;
  NumeroBairroEndereco: number | null;
  NumeroLogadouroEndereco: number | null;
  CodigoEmpresaPessoa: number | null;
  NomeEmpresaPessoa: string | null;
  TipoEnderecoEmpresa: number | null;
};